Add ability to redirect znoyder outptut to file
This introduces two cli parameters, log_mode and log_file. The first
gives the user the ability to choose whether to write the ouptut to
stderr, to a file or to both. In case file is selected, the log_file
parameter selects the path to write to. This parameters can be specified
either as general znoyder commands or as specific options for a command.